requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
databases==0.7.0
exceptiongroup==1.1.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
packaging==23.0
pluggy==1.0.0
pydantic==1.10.5
pypika-tortoise==0.1.6
pytest==7.2.1
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
tomli==2.0.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
